+++ aead/api/lc_ascon_lightweight.h
@@ -17,54 +17,54 @@
  * DAMAGE.
  */
 
-#ifndef LC_ASCON_KECCAK_H
-#define LC_ASCON_KECCAK_H
+#ifndef LC_ASCON_LIGHTWEIGHT_H
+#define LC_ASCON_LIGHTWEIGHT_H
 
 #include "lc_ascon_aead.h"
-#include "lc_sha3.h"
+#include "lc_ascon_hash.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#define LC_AK_STATE_SIZE (LC_SHA3_STATE_SIZE + LC_ASCON_ALIGNMENT)
-#define LC_AK_CTX_SIZE(x)                                                      \
+#define LC_AL_STATE_SIZE (LC_ASCON_HASH_STATE_SIZE + LC_ASCON_ALIGNMENT)
+#define LC_AL_CTX_SIZE(x)                                                      \
 	(sizeof(struct lc_aead) + sizeof(struct lc_ascon_cryptor) +            \
-	 LC_AK_STATE_SIZE)
+	 LC_AL_STATE_SIZE)
 
 /**
- * @brief Allocate Ascon Keccak cryptor context on heap
+ * @brief Allocate Ascon Lightweight cryptor context on heap
  *
- * NOTE: This is defined for lc_sha3_512 and lc_sha3_256.
+ * NOTE: This is defined for lc_ascon_128*.
  *
  * @param [in] hash Hash implementation of type struct hash used for the
  *		    Ascon-Keccak algorithm
- * @param [out] ctx Allocated Ascon-Keccak cryptor context
+ * @param [out] ctx Allocated Ascon lightweight cryptor context
  *
  * @return 0 on success, < 0 on error
  */
-int lc_ak_alloc(const struct lc_hash *hash, struct lc_aead_ctx **ctx);
+int lc_al_alloc(const struct lc_hash *hash, struct lc_aead_ctx **ctx);
 
 /**
- * @brief Allocate stack memory for the Ascon-Keccak cryptor context
+ * @brief Allocate stack memory for the Ascon lightweight cryptor context
  *
- * NOTE: This is defined for lc_sha3_512 and lc_sha3_256.
+ * NOTE: This is defined for lc_ascon_128* as of now.
  *
  * @param [in] name Name of the stack variable
  * @param [in] hash Hash implementation of type struct hash used for the
- *		    Ascon-Keccak algorithm
+ *		    Ascon lightweight algorithm
  */
-#define LC_AK_CTX_ON_STACK(name, hash)                                              \
+#define LC_AL_CTX_ON_STACK(name, hash)                                              \
 	_Pragma("GCC diagnostic push")                                              \
 		_Pragma("GCC diagnostic ignored \"-Wvla\"") _Pragma(                \
 			"GCC diagnostic ignored \"-Wdeclaration-after-statement\"") \
 			LC_ALIGNED_BUFFER(name##_ctx_buf,                           \
-					  LC_AK_CTX_SIZE(hash),                     \
-					  LC_ASCON_ALIGNMENT);               \
+					  LC_AL_CTX_SIZE(hash),                     \
+					  LC_ASCON_ALIGNMENT);                      \
 	struct lc_aead_ctx *name = (struct lc_aead_ctx *)name##_ctx_buf;            \
 	LC_ASCON_SET_CTX(name, hash);                                               \
 	struct lc_ascon_cryptor *__name_ascon_crypto = name->aead_state;            \
-	__name_ascon_crypto->statesize = LC_SHA3_STATE_SIZE;                        \
+	__name_ascon_crypto->statesize = LC_ASCON_HASH_STATE_SIZE;                  \
 	_Pragma("GCC diagnostic pop")
 /* invocation of lc_ak_zero_free(name); not needed */
 
@@ -72,4 +72,4 @@ int lc_ak_alloc(const struct lc_hash *hash, struct lc_aead_ctx **ctx);
 }
 #endif
 
-#endif /* LC_ASCON_KECCAK_H */
+#endif /* LC_ASCON_LIGHTWEIGHT_H */