+++ hash/src/sha3_avx512.c
@@ -115,8 +115,10 @@ static void keccak_avx512_squeeze(void *_state, uint8_t *digest)
 	LC_FPU_DISABLE;
 }
 
-static void keccak_avx512_permutation(void *state)
+static void keccak_avx512_permutation(void *state, unsigned int rounds)
 {
+	(void)rounds;
+
 	LC_FPU_ENABLE;
 	KeccakP1600_AVX512_Permute_24rounds(state);
 	LC_FPU_DISABLE;