+++ CHANGES.md
@@ -6,6 +6,8 @@ Changes 0.10.0
 
 * update AEAD: add lc_aead_enc|dec_init and change all AEAD algo's tag calculation to now perform MAC(AAD || ciphertext) instead of MAC(ciphertext || AAD) - this brings it in line with all AEAD algorithms
 
+* enhancement: add Ascon AEAD 128 and 128b
+
 Changes 0.9.2
 
 * fix: update "reduce memory footprint of Keccak state" to handle big-endian systems